refactoring from andi
