+++ src/main/java/module-info.java
@@ -9,5 +9,6 @@ module at.ac.fhcampuswien.fhmdb {
     opens at.ac.fhcampuswien.fhmdb to javafx.fxml;
     exports at.ac.fhcampuswien.fhmdb;
     exports at.ac.fhcampuswien.fhmdb.models;
+    exports at.ac.fhcampuswien.fhmdb.utils;
 
 }
